+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.IntuneDeviceAndAppManagementAssignmentFilter.Tests.ps1
@@ -47,7 +47,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
             }
             Mock -CommandName Remove-MgDeviceManagementAssignmentFilter -MockWith {
             }
-
+        }
             # Test contexts
             Context -Name "When the Filter doesn't already exist" -Fixture {
                 BeforeAll {
@@ -197,7 +197,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                 }
             }
         }
-    }
+    
 }
 
 Invoke-Command -ScriptBlock $Global:DscHelper.CleanupScript -NoNewScope